IE Explorer mock browser at MVP - icons buttons janky done
all rows finished:
- links
- dropdowns
- address branch
- icons - which have the hover effect - lots of time spent on creating, editing, resizing icons. happy with the jankiness of it.

index of section icons and spacing working easily

Refresh, Stop, Fullscreen, Back, and Forward use the new system which has a base .icon class and BEM syntax for differing icon sizes and positionings

index.html is now a combination and refactoring of tree-factor.html for the procedural generation and browser.html for the fake ie 4

TODO:
  - refactor browser.css most icons are using an older system with lots of duplicated css

  - go back to bees.html/js/css and add the:
    - buttons
      - the daisies in the garden will make changes to the text frame
      - or trigger other things on the page??
    - links - something on the page - maybe a daisy, maybe not will take you back to the index of page

the rest of the changes have been due to either other experimentation, proof-of-concepts, failed explorations and aren't worthy to mention.

other than those the other changes have been in session with Tabi and Myra. we spend 3 - 5 hrs working on what we have been calling 'interstitials' in these glades. the interstitials:
  - animalrave.html/js/css
  - bees.html/js/css
  - elfs a folder to be added later to the next commit
  scenes between the nodes somewhere